api/requirements.txt
Flask==2.2.2
Flask-Cors==3.0.10
python-dotenv==0.21.0
together
pyarrow==18.1.0
numpy==1.24.0